read new file

+++ src/zecora.h
@@ -29,7 +29,7 @@
 
 void prepareFrameBuffer();
 void createScratch();
-int openFile(char* filename);
+long openFile(char* filename);
 void jump(char* command);
 void createScreen(int rows, int cols);
 void readInput(int cols);
